Page numbers after word characters now icrement by sign. Fixed addition of "-" before negative numbers.
"page-4" and "page4" are ambiguous and both are common enough. Both are handled as they were "4" not e.g. "-4". This means incrementing "page-4" goes to "page-5", and  "page4" goes to "page5". Behavior with the usual format (page=45, /45) is not changed.

The bug causing and extra "-" added before e.g. "-5" is fixed.